The Unofficial Planco Dictionary API
This script takes the Official Planco Dictionary (https://www.planetcoaster.com/en-gb/official-planco-dictionary) and converts it to an array of word objects.

Because there isn't a Official Planco Dictionary API, the only way to create a Eng/Plc dictionary application is to manually input the data for each word (2249 words as of 9/18/19, making 8,996 items to be copied!). But after copying and pasting each element of about 450 words, I started to see how long and tedious this process was actually becoming.  So , I wrote this!

Open the link and paste it into the console. Done.